Pass C include paths to runCCdep
In case the C sources rely on other files (usually headers) in other
directories, then "c-libpaths" is used when building the target.
OTOH, this will not help during the C dependency discovery, since the
compiler is invoked without those, and thus it will fail.

Hence, use the include paths for a target when running its C dependency
discovery.