Hopefully fix C dependencies for OCaml linking
Nodes for C sources are added to the DAGs of each target using them,
without creating any edge between these nodes and the targets.
The result is that an OCaml linking task does not wait for all the C
builds to finish, resulting in missing files halting the linking.

Hence, set the proper edge between the CompileC node and the LinkTarget
one, hopefully fixing the dependency.  This also means the hack of
waiting for the .o files is not needed anymore.